LinearMap: k-v lifetime fix
Given that keys and values may not be default constructible and out map
will initially hold invalid values, let's use a memory block and cast it
to the actual type when used. This ensures we don't have to default
construct them. Also, given that there is no real object, we need to
make sure we call destructors when objects are erased, and also when the
map is destructed.