Add ReadWriteLock<T, M>
The ReadWriteLock differs from a ReasourceLock in that it can be locked
multiple times by multiple readers from differents (or the same) threads
because they do not mutate the state of the resource at any point in
time, making it safe to use it concurrently for reading from multiple
threads.

On the other hand, write locks are exclusive and cannot be held while
any other thread (or the same thread) holds another read or write lock.